Add shading option to plot_model_field()
When using plt.tripcolor(), shading can be 'flat' (required when data are provided on elements, or averages data on points), or 'gouraud' (data defined and plotted on points). Here, we use 'gouraud' as default to ensure the range of data matches exactly when data are provided on points. This is more consistent with the MATLAB plotting implementation. When data are defined on elements, 'flat' is required. This is adjusted internally and a warning generated.