Add plot_model_bc() and extract model data on elements (Issue #37 and #32) (#42)
* Extract layer from 3D model w/ data on elements

Add functionality to extract data defined on elements from a 3D model. This addresses issue #32. Updated to return select_indices as well to identify the vertex/element numbers associated with the extracted layer

* Add shading option to plot_model_field()

When using plt.tripcolor(), shading can be 'flat' (required when data are provided on elements, or averages data on points), or 'gouraud' (data defined and plotted on points). Here, we use 'gouraud' as default to ensure the range of data matches exactly when data are provided on points. This is more consistent with the MATLAB plotting implementation. When data are defined on elements, 'flat' is required. This is adjusted internally and a warning generated.

* Update shading option to be more user interpretable

* Add plot_model_bc()