Add support for time-varying input fields
Some input fields have an additional row (md.mesh.numberofvertices + 1 or md.mesh.numberofelements + 1) that includes a timestep definition for each column. This needs to be removed for plotting so that the size of the field matches the number of points/elements. Added this correction and associated warning message. Addresses Issue #47