Updates to plotting func & rename build -> param in model sub-class doc-strings (#51)
* Replace 'build' with 'param' in doc strings

* Fix plotting bug

* Add plot_model_ts() for 1D timeseries plotting

* Minor updates to some plotting funcs

* Add support for pd.Dataframe variable map

1) variable_map can now be passed to the function as a pandas Dataframe or a *.csv file. 2) If a requested variable exists in the model, but is empty (e.g. if there is no md.geometry.bed defined in the model), it simply prints a warning and continues. The printed warning includes the group (and sub-group if issm_subgroup is not empty).

* Add support for time-varying input fields

Some input fields have an additional row (md.mesh.numberofvertices + 1 or md.mesh.numberofelements + 1) that includes a timestep definition for each column. This needs to be removed for plotting so that the size of the field matches the number of points/elements. Added this correction and associated warning message. Addresses Issue #47