Merge 160925_lb_model_io -- Fix #60, add checks for C++ Wrappers/ISSM install. (#66)
* Fix Issue #60 to serialise boolean instances

* Implement lazy-loading for C++ wrappers

When no wrappers were installed (.so libraries did not exist), an error occurred during the import of utils.wrappers as the load_issm_wrapper() decorator was being executed at import. This new implementation runs a _load_func() function when a given wrapper is called. This loads the function (if it hasn't already been loaded). This works with local ISSM install with/without wrappers. To be tested on gadi.

* Add checks for wrapper-dependent functions

Some functions require C++ wrappers for full functionality. Add checks with warnings (and return empty for manual definition) or errors (where hard requirement exists) as necessary. Loosen check for '_python.so' in $ISSM_DIR/lib to look for *_python.* to allow for different system versions (e.g. dylib on some Mac systems).

* Update ISSM_DIR / wrapper checks