Fix empty list handling with model IO
When a list is empty, it gets written to/from file as a list with an empty string. This causes issues down-the-road for classes/variables that are expecting a list to be empty. Here, we add a special case to convert empty lists back to empty lists when loading models.