Add additional functionality required for Antarctic-wide model setup (#85)
* Correct dimensions

* Add parameterize() and contour_envelope()

* Add [existing ISSM] extract() functionality

* Add deactivate_all() function; verbose bool to int

Change bool to int for default verbose usage

* Loosen checks for ISSM Directory

* Add gadi cluster

* Minor fixes

Fix m1qn3 inversion checks and default inversion settings; Update initialization marshalling; Correct typo in rifts

* Fix empty list handling with model IO

When a list is empty, it gets written to/from file as a list with an empty string. This causes issues down-the-road for classes/variables that are expecting a list to be empty. Here, we add a special case to convert empty lists back to empty lists when loading models.

* no_log option for solve/load_results_from_cluster

* Add new functions to tools

* Add data.interp and remove data.datapool

Datapool functionality is being moved outside pyISSM, so remove this. Add data.interp for interpolating model data

* Update issm_scp_in for missing errlog

* Update formatting of outlog

* Add kill_icebergs()

* Update interpolation routines